fix(Tree): drag preview support ReactNode type
